backend/requirements/common.txt
#requrements/common.txt
#python==3.5.3
Django==1.11.5
mysqlclient==1.3.12
djangorestframework==3.6.4
django-cors-headers==2.1.0
django-pagination==1.0.7